events code break down into separate folders
