+++ src/events/events.schema.ts
@@ -1,4 +1,4 @@
-import { boolean, number, object, string, array, mixed, date } from 'yup';
+import { number, object, string } from 'yup';
 
 export const trackSchema = object().shape({
   eventName: string().required(),
@@ -7,76 +7,4 @@ export const trackSchema = object().shape({
   dataFields: object(),
   campaignId: number(),
   templateId: number()
-});
-
-export const eventRequestSchema = object().shape({
-  messageId: string().required(),
-  clickedUrl: string(),
-  messageContext: object().shape({
-    saveToInbox: boolean(),
-    silentInbox: boolean(),
-    location: string()
-  }),
-  closeAction: string(),
-  deviceInfo: object()
-    .shape({
-      deviceId: string().required(),
-      platform: string().required(),
-      appPackageName: string().required()
-    })
-    .required(),
-  inboxSessionId: string(),
-  createdAt: number()
-});
-
-export const trackEmbeddedMessageSchema = object().shape({
-  metadata: object().shape({
-    messageId: string(),
-    campaignId: number(),
-    isProof: boolean(),
-    placementId: number()
-  }),
-  elements: object().shape({
-    title: string(),
-    body: string(),
-    mediaUrl: string(),
-    buttons: array().of(object().shape({
-      id: string(),
-      title: string(),
-      action: object().shape({
-        type: string(),
-        data: string()
-      })
-    })),
-    text: array().of(object().shape({
-      id: string(),
-      text: string()
-    })),
-    defaultAction: array().of(object().shape({
-      type: string(),
-      data: string()
-    })),
-  }),
-  payload: array().of(mixed())
-});
-
-export const trackEmbeddedMessageClickSchema = object().shape({
-  messageId: string(),
-  buttonIdentifier: string(),
-  targetUrl: string(),
-  deviceInfo: object().shape({
-    appPackageName: string(),
-  }),
-});
-
-export const trackEmbeddedSessionSchema = object().shape({
-  start: date(),
-  end: date(),
-  placementId: string(),
-  impressions: array().of(object().shape({
-    messageId: string(),
-    displayCount: number(),
-    duration: number()
-  })),
-  id: string(),
 });
