+++ src/events/events.ts
@@ -1,20 +1,7 @@
 import { baseIterableRequest } from '../request';
-import {
-  InAppEventRequestParams,
-  InAppTrackRequestParams,
-  IEmbeddedMessage,
-  IEmbeddedMessageMetadata,
-  IEmbeddedSession
-} from './types';
+import { InAppTrackRequestParams } from './in-app/types';
 import { IterableResponse } from '../types';
-import { WEB_PLATFORM } from '../constants';
-import {
-  eventRequestSchema,
-  trackSchema,
-  trackEmbeddedMessageSchema,
-  trackEmbeddedMessageClickSchema,
-  trackEmbeddedSessionSchema
-} from './events.schema';
+import { trackSchema } from './events.schema';
 import { EndPoints } from './consts';
 
 export const track = (payload: InAppTrackRequestParams) => {
@@ -28,192 +15,6 @@ export const track = (payload: InAppTrackRequestParams) => {
     data: payload,
     validation: {
       data: trackSchema
-    }
-  });
-};
-
-export const trackEmbeddedMessageReceived = (payload: IEmbeddedMessage) => {
-  return baseIterableRequest<IterableResponse>({
-    method: 'POST',
-    url: EndPoints.msg_received_event_track,
-    data: payload,
-    validation: {
-      data: trackEmbeddedMessageSchema
-    }
-  });
-};
-
-export const trackEmbeddedMessageClick = (
-  payload: IEmbeddedMessageMetadata,
-  buttonIdentifier: string,
-  clickedUrl: string,
-  appPackageName: string
-) => {
-  return baseIterableRequest<IterableResponse>({
-    method: 'POST',
-    url: EndPoints.msg_click_event_track,
-    data: {
-      messageId: payload.messageId,
-      buttonIdentifier: buttonIdentifier,
-      targetUrl: clickedUrl,
-      deviceInfo: {
-        appPackageName: appPackageName
-      }
-    },
-    validation: {
-      data: trackEmbeddedMessageClickSchema
-    }
-  });
-};
-
-export const trackEmbeddedSession = (payload: IEmbeddedSession) => {
-  return baseIterableRequest<IterableResponse>({
-    method: 'POST',
-    url: EndPoints.msg_impression_event_track,
-    data: payload,
-    validation: {
-      data: trackEmbeddedSessionSchema
-    }
-  });
-};
-
-export const trackInAppClose = (payload: InAppEventRequestParams) => {
-  /* a customer could potentially send these up if they're not using TypeScript */
-  delete (payload as any).userId;
-  delete (payload as any).email;
-
-  return baseIterableRequest<IterableResponse>({
-    method: 'POST',
-    url: EndPoints.track_app_close,
-    data: {
-      ...payload,
-      deviceInfo: {
-        ...payload.deviceInfo,
-        platform: WEB_PLATFORM,
-        deviceId: global.navigator.userAgent || ''
-      }
-    },
-    validation: {
-      data: eventRequestSchema
-    }
-  });
-};
-
-export const trackInAppOpen = (
-  payload: Omit<
-    InAppEventRequestParams,
-    'clickedUrl' | 'inboxSessionId' | 'closeAction'
-  >
-) => {
-  /* a customer could potentially send these up if they're not using TypeScript */
-  delete (payload as any).userId;
-  delete (payload as any).email;
-
-  return baseIterableRequest<IterableResponse>({
-    method: 'POST',
-    url: EndPoints.track_app_open,
-    data: {
-      ...payload,
-      deviceInfo: {
-        ...payload.deviceInfo,
-        platform: WEB_PLATFORM,
-        deviceId: global.navigator.userAgent || ''
-      }
-    },
-    validation: {
-      data: eventRequestSchema.omit([
-        'clickedUrl',
-        'inboxSessionId',
-        'closeAction'
-      ])
-    }
-  });
-};
-
-export const trackInAppClick = (
-  payload: Omit<InAppEventRequestParams, 'inboxSessionId' | 'closeAction'>,
-  sendBeacon = false
-) => {
-  /* a customer could potentially send these up if they're not using TypeScript */
-  delete (payload as any).userId;
-  delete (payload as any).email;
-
-  return baseIterableRequest<IterableResponse>({
-    method: 'POST',
-    url: EndPoints.track_app_click,
-    sendBeacon,
-    data: {
-      ...payload,
-      deviceInfo: {
-        ...payload.deviceInfo,
-        platform: WEB_PLATFORM,
-        deviceId: global.navigator.userAgent || ''
-      }
-    },
-    validation: {
-      data: eventRequestSchema.omit(['inboxSessionId', 'closeAction'])
-    }
-  });
-};
-
-export const trackInAppDelivery = (
-  payload: Omit<
-    InAppEventRequestParams,
-    'clickedUrl' | 'closeAction' | 'inboxSessionId'
-  >
-) => {
-  /* a customer could potentially send these up if they're not using TypeScript */
-  delete (payload as any).userId;
-  delete (payload as any).email;
-
-  return baseIterableRequest<IterableResponse>({
-    method: 'POST',
-    url: EndPoints.track_app_delivery,
-    data: {
-      ...payload,
-      deviceInfo: {
-        ...payload.deviceInfo,
-        platform: WEB_PLATFORM,
-        deviceId: global.navigator.userAgent || ''
-      }
-    },
-    validation: {
-      data: eventRequestSchema.omit([
-        'clickedUrl',
-        'inboxSessionId',
-        'closeAction'
-      ])
-    }
-  });
-};
-
-export const trackInAppConsume = (
-  payload: Omit<
-    InAppEventRequestParams,
-    'clickedUrl' | 'closeAction' | 'inboxSessionId'
-  >
-) => {
-  /* a customer could potentially send these up if they're not using TypeScript */
-  delete (payload as any).userId;
-  delete (payload as any).email;
-
-  return baseIterableRequest<IterableResponse>({
-    method: 'POST',
-    url: EndPoints.track_app_consume,
-    data: {
-      ...payload,
-      deviceInfo: {
-        ...payload.deviceInfo,
-        platform: WEB_PLATFORM,
-        deviceId: global.navigator.userAgent || ''
-      }
-    },
-    validation: {
-      data: eventRequestSchema.omit([
-        'clickedUrl',
-        'inboxSessionId',
-        'closeAction'
-      ])
     }
   });
 };
