+++ src/events/in-app/events.ts
@@ -1,0 +1,147 @@
+import { baseIterableRequest } from '../../request';
+import { InAppEventRequestParams } from './types';
+import { IterableResponse } from '../../types';
+import { WEB_PLATFORM } from '../../constants';
+import { eventRequestSchema } from './events.schema';
+import { EndPoints } from '../consts';
+
+export const trackInAppClose = (payload: InAppEventRequestParams) => {
+  /* a customer could potentially send these up if they're not using TypeScript */
+  delete (payload as any).userId;
+  delete (payload as any).email;
+
+  return baseIterableRequest<IterableResponse>({
+    method: 'POST',
+    url: EndPoints.track_app_close,
+    data: {
+      ...payload,
+      deviceInfo: {
+        ...payload.deviceInfo,
+        platform: WEB_PLATFORM,
+        deviceId: global.navigator.userAgent || ''
+      }
+    },
+    validation: {
+      data: eventRequestSchema
+    }
+  });
+};
+
+export const trackInAppOpen = (
+  payload: Omit<
+    InAppEventRequestParams,
+    'clickedUrl' | 'inboxSessionId' | 'closeAction'
+  >
+) => {
+  /* a customer could potentially send these up if they're not using TypeScript */
+  delete (payload as any).userId;
+  delete (payload as any).email;
+
+  return baseIterableRequest<IterableResponse>({
+    method: 'POST',
+    url: EndPoints.track_app_open,
+    data: {
+      ...payload,
+      deviceInfo: {
+        ...payload.deviceInfo,
+        platform: WEB_PLATFORM,
+        deviceId: global.navigator.userAgent || ''
+      }
+    },
+    validation: {
+      data: eventRequestSchema.omit([
+        'clickedUrl',
+        'inboxSessionId',
+        'closeAction'
+      ])
+    }
+  });
+};
+
+export const trackInAppClick = (
+  payload: Omit<InAppEventRequestParams, 'inboxSessionId' | 'closeAction'>,
+  sendBeacon = false
+) => {
+  /* a customer could potentially send these up if they're not using TypeScript */
+  delete (payload as any).userId;
+  delete (payload as any).email;
+
+  return baseIterableRequest<IterableResponse>({
+    method: 'POST',
+    url: EndPoints.track_app_click,
+    sendBeacon,
+    data: {
+      ...payload,
+      deviceInfo: {
+        ...payload.deviceInfo,
+        platform: WEB_PLATFORM,
+        deviceId: global.navigator.userAgent || ''
+      }
+    },
+    validation: {
+      data: eventRequestSchema.omit(['inboxSessionId', 'closeAction'])
+    }
+  });
+};
+
+export const trackInAppDelivery = (
+  payload: Omit<
+    InAppEventRequestParams,
+    'clickedUrl' | 'closeAction' | 'inboxSessionId'
+  >
+) => {
+  /* a customer could potentially send these up if they're not using TypeScript */
+  delete (payload as any).userId;
+  delete (payload as any).email;
+
+  return baseIterableRequest<IterableResponse>({
+    method: 'POST',
+    url: EndPoints.track_app_delivery,
+    data: {
+      ...payload,
+      deviceInfo: {
+        ...payload.deviceInfo,
+        platform: WEB_PLATFORM,
+        deviceId: global.navigator.userAgent || ''
+      }
+    },
+    validation: {
+      data: eventRequestSchema.omit([
+        'clickedUrl',
+        'inboxSessionId',
+        'closeAction'
+      ])
+    }
+  });
+};
+
+export const trackInAppConsume = (
+  payload: Omit<
+    InAppEventRequestParams,
+    'clickedUrl' | 'closeAction' | 'inboxSessionId'
+  >
+) => {
+  /* a customer could potentially send these up if they're not using TypeScript */
+  delete (payload as any).userId;
+  delete (payload as any).email;
+
+  return baseIterableRequest<IterableResponse>({
+    method: 'POST',
+    url: EndPoints.track_app_consume,
+    data: {
+      ...payload,
+      deviceInfo: {
+        ...payload.deviceInfo,
+        platform: WEB_PLATFORM,
+        deviceId: global.navigator.userAgent || ''
+      }
+    },
+    validation: {
+      data: eventRequestSchema.omit([
+        'clickedUrl',
+        'inboxSessionId',
+        'closeAction'
+      ])
+    }
+  });
+};
