+++ src/events/index.ts
@@ -1,2 +1,5 @@
-export * from './types';
 export * from './events';
+export * from './in-app/events';
+export * from './in-app/types';
+export * from './embedded/events';
+export * from './embedded/types';
